feat(si-fs): repurpose `install.sh` script to install `si-fs`
This change updates the pre-existing `install.sh` script (found under
the auth portal's `public/` directory) to target the downloading and
installation of a `bin/si-fs` binary from the artifact repository.

The help usage is as follows:

```
install.sh

Installs a binary release of si-fs for supported platforms

USAGE:
    install.sh [OPTIONS] [--]

OPTIONS:
    -h, --help                Prints help information
    -d, --destination=<DEST>  Destination directory for installation
                              [default: /home/fnichol/.local/bin]
    -p, --platform=<PLATFORM> Platform type to install
                              [examples: linux-x86_64, darwin-aarch64]
                              [default: linux-x86_64]
    -V, --version=<VERSION>   Release version to install
                              [examples: stable,
                              20250218.210911.0-sha.bda1ce6ea]
                              [default: stable]

EXAMPLES:
    # Installs the latest release into `$HOME/bin`
    install.sh

    # Installs the latest release for all users under `/usr/local/bin`
    sudo install.sh

    # Installs an old release into a temp directory
    install.sh -V 20250214.193652.0-sha.0f9972a53 -d /tmp

```

The default will be to download the current `stable` version for the
target platform (currently only `linux-x86_64` and `linux-aarch64` are
supported) and install it under one of the following locations (first
path that is present in the user's `$PATH` wins):

- `$HOME/.local/bin`
- `$HOME/bin`
- Falls back to `$HOME/.si-fs/bin` if none of the above are found in
  `$PATH`

If a `root` user (that is `uid == 0`) is installing, the default
destination will be `/usr/local/bin`.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>