feat(dal,sdf): Auto-enqueue update functions for impacted components
For now, this is behind a feature flag as we work out the various permutations and behavior we want to see.  

When an attribute value is updated for a component, and the component has a resource AND an update function exists, enqueue it. Then, build a dependency graph, and if any downstream components with resources also have update functions, enqueue them as well.