Make module sync rely on local cache only
This commit makes the module sync route rely on the local cache only
without requiring queries to the module index. It also removes the
tab to install assets since dynamic installation is now in play. Since
the tab is removed and there's only one tab for most users of SI, the
tab group will be disabled and the pill counter will move to the title
of the assets panel.

Signed-off-by: Nick Gerace <nick@systeminit.com>