feat(clover): Change how JSON Numbers are represented in assets
We are going to take JSON Numbers as Float. This will allow us to support their precision. We need to ensure we add the setFloat intrinsic for this

We have the new intrinsics for resourcePayloadToValue and NormalizeToArray so we need to use those in the pkgSpecs

We are also changing to use the AWS Colour pallet - #FF9900