Add "migration-mode" to remove workspace snapshots unused by any active change sets
While we do evict snapshots when a change set pointer moves, we don't do
anything when a change set's status is no longer one of the "active"
statuses. We also do not currently allow setting the status of a change
set back to one of the "active" statuses once it has been moved out of
being "active".

Because a change set is forever "inactive" once it has become inactive,
it is not necessary to keep the snapshot around, unless it is also being
used by an "active" change set.

This new migration-mode (`garbageCollectSnapshots`) queries all "active"
change sets to gather their workspace snapshot addresses, and all of the
workspace snapshot addresses in the LayerDb. Any workspace snapshot that
is at least an hour old, and is _NOT_ referenced by an "active" change
set is removed. We only consider snapshots older than an hour to avoid
race conditions where a change set might have been created or modified
after we queried the `change_set_pointers` table, and before when we
query the `workspace_snapshots` table. Because the tables are in
separate databases, we can't rely on normal transactional integrity.