Operation Verinuke
This commit adds the "verinuke" in addition to a "pause-resume" stream
wrapper to halt verideath or, at least, delay the inevitable.
The stream wrapper is flagged by an application-level config option and
is enabled by default.

Let's start with verinuke: we will now SIGTERM the veritech process,
giving a chance for cyclone/firecracker to not have zombies if we ever
see a missing heartbeat. This is a temporary measure until we figure
out what is causing our local NATS client(s) (cloned from the main
veritech process) to die.

Now for the stream wrapper: it allows for a few things...

- "Pause" or "Resume" the inflow of data by unsubscribing and backing
  off until re-subscribing (much like an HTTP server shutoff valve)
- Handle "MissingHeartbeat" errors with its own backoff timer (and for
  now, with verinuke)
- Initialize and manage the subscription lifecycle via state machine

The first and only "pause" location is located in the sole veritech
handler for the core app. It triggers when the PoolNoodle pool has been
exhausted. Why here? We believe that this error causes a hard-hitting
NACK loop that overwhelms our NATS client(s) locally (not over the
network).

When the feature is enabled, you can tweak the pause duration and
reconnect backoff duration values at the application config level.

As part of these changes, some metrics have been tweaked as well as
log levels for existing naxum and NATS-related tracing.

The "five failed message" circuit-breaker has been removed from naxum
since it puts the naxum application user in an unrecoverable state and
is pushed too far down into the library domain. Note that this has been
removed from _both_ veritech naxum apps and not just the one disabled
by default that's been added to this commit.

Finally, we now add a customizable graceful shutdown timeout to veritech
as well as reset all metrics related to event callbacks for new NATS
clients. The default timeout has moved from 6 hours to 15 seconds,
which is a massive change, but it is now override-able.

Co-authored-by: Brit Myers <brit@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>