feat: si-split-graph
A new crate, lib/si-split-graph, containing an initial, generic
implementation of a merkle tree hashed graph which can be split into
multiple "subgraphs" but allows edges between nodes in different
subgraphs. Calculates updates, performs them, and handles the "ordering"
concept internally.

This is not yet integrated with the si "dal", and will likely require
more changes before that is possible.