bug: improve removeEmtpy in clover assets
We've added sockets that are more deeply nested in some schemas than we
have before. The side effect is that we can fail to remove some empty
objects or arrays from the tree, resulting in invalid CloudFormation.

This PR fixes it by rewriting that function to handle deeply nested
cases, and to remove its use of recursion (so we don't blow the stack on
huge objects).