feat(si-api-test): Add a new benchmarking test to trigger a stampede of func executions bypassing DVU
To aid in testing/measuring the performance of function execution, this new test takes advantage of the ability to test function executions, where SDF dispatches functions directly to veritech without a Pinga Job.  

I've also included a helpful python script we've been using to execute these tests and some instructions on how to use it!