feat: move the removeEmtpy function to a jsr module
This moves the 'removeEmtpy' function, which we are reusing in lots of
places, into a small JSR package. Then we can import that package with
the major version number specified, and get automatic upgrades without
having to republish all the assets in the universe.

See also https://jsr.io/@systeminit/remove-empty

For now, we plumb this through into a sandbox namespace called 'exLib'.
That just imports the function and then exposes it. We will still
publish updates to JSR, and then bump the version number in lang-js,
which will allow us to not worry about JSR caching. We will take an
opportunity in the future to move toward calling the JSR deps directly
once we figure out caching in veritech/cyclone/lang-js.