Add veritech heartbeat app, naxum metrics and more
This commit adds a heartbeat app with force reconnection abilities to
veritech, adds naxum metrics and tunes metrics and telemetry everywhere.

Many of the changes in this commit come from the testing grounds PR
(#5566), but not all changes were included. The "pause-resume" stream
wrapper and client hot-swapper changes are either not ready for
production use or are not tactics we should employ for one reason or
another.

For veritech, the new heartbeat app is disabled in tests (by default)
and enabled in all other scenarios (by default). The heartbeat app
publishes messages via core NATS on a cadence and works with subject
prefixes. It uses the instance ID to determine the destination subject.

Metrics have been added to track pool exhaustion, publisher stats,
client stats, handlers doing work and more.

Upon shutdown, we will now explicitly cancel the shutdown token if the
stream has closed for the core app so that the kill app also closes.

For naxum, metrics have been added for the core service and in
middleware. Tracing has been tuned to be as noiseless as possible while
making use of metrics where possible. These changes include tracking
failures for "double ack failure" and standardizing maintain progress
task logging.

The "ciruitbreaker" with 4 max messages has been removed as a result of
these changes. Not only is it an anti-pattern for handling NACK'd
messages, but it also masks over potential poison pill messages for
consumers who allow for more than 4 re-deliveries.

For NATS event callbacks, all trace logging is gone and now everything
flows through metrics.

Finally, the graceful shutdown time is now tunable for veritech and its
default has moved from 6 hours to 20 minutes.

Signed-off-by: Nick Gerace <nick@systeminit.com>