feat: add a `SIGQUIT` listener that writes a Tokio tasks report file
This change adds a signal listener to all Rust `bin/*` binaries, that
upon receiving a `SIGQUIT` signal, will write a "process report" file to
the temp file system, named with the service name and second-resolution
timestamp. This report currently prints a trace for every currently
spawned task on the main Tokio runtime.

The file looks something like this (appologies for the Markdown escapes,
it's the only way to render in a Git commit message *and* GitHub PR
message without being dropped):

```
\# veritech Report (2025-03-12T16:59:18Z)

\## Tokio Task Traces

\### Task 0 Trace
\```
╼ <core::pin::Pin<P> as core::future::future::Future>::poll at /nix/store/yxdqfnnr48j8m5h00806i4l2vqi4fiyr-rust-default-1.84.1/lib/rustlib/src/rust/library/core/src/future/future.rs:124:9
  └╼ <tower::buffer::worker::Worker<T,Request> as core::future::future::Future>::poll at ./third-party/rust/tower-0.4.13.crate/src/buffer/worker.rs:181:26
     └╼ tower::buffer::worker::Worker<T,Request>::poll_next_msg at ./third-party/rust/tower-0.4.13.crate/src/buffer/worker.rs:122:38
        └╼ tokio::sync::mpsc::unbounded::UnboundedReceiver<T>::poll_recv at ./third-party/rust/tokio-1.44.0.crate/src/sync/mpsc/unbounded.rs:432:9
           └╼ tokio::sync::mpsc::chan::Rx<T,S>::recv at ./third-party/rust/tokio-1.44.0.crate/src/sync/mpsc/chan.rs:292:16
\```

\### Task 1 Trace
\```
╼ <core::pin::Pin<P> as core::future::future::Future>::poll at /nix/store/yxdqfnnr48j8m5h00806i4l2vqi4fiyr-rust-default-1.84.1/lib/rustlib/src/rust/library/core/src/future/future.rs:124:9
  └╼ opentelemetry_sdk::trace::span_processor::BatchSpanProcessor<R>::new::{{closure}} at ./third-party/rust/opentelemetry_sdk-0.26.0.crate/src/trace/span_processor.rs:474:37
     └╼ opentelemetry_sdk::trace::span_processor::BatchSpanProcessorInternal<R>::run::{{closure}} at ./third-party/rust/opentelemetry_sdk-0.26.0.crate/src/trace/span_processor.rs:428:13
        └╼ <futures_util::future::poll_fn::PollFn<F> as core::future::future::Future>::poll at ./third-party/rust/futures-util-0.3.31.crate/src/future/poll_fn.rs:56:9
           └╼ opentelemetry_sdk::trace::span_processor::BatchSpanProcessorInternal<R>::run::{{closure}}::{{closure}} at ./third-party/rust/futures-util-0.3.31.crate/src/async_await/select_mod.rs:323:13
              └╼ core::ops::function::impls::<impl core::ops::function::FnMut<A> for &mut F>::call_mut at /nix/store/yxdqfnnr48j8m5h00806i4l2vqi4fiyr-rust-default-1.84.1/lib/rustlib/src/rust/library/core/src/ops/function.rs:294:13
 
...snip...
```

The file created is determined by using the `TMPDIR`/`TMP`/`TEMP`
environment variables first, falling back to `/tmp` and would be named
something like:

````
/tmp/veritech.report.2025-03-12T16:59:18Z.md
````

Note that in our Nix development environment, it's more likely that the
file would be something like:

````
/tmp/nix-shell.LnbroP/nix-shell.Exe6tX/veritech.report.2025-03-12T16:59:18Z.md
````

The main process will emit an `info` line to confirm that a report file
is being written out and have its filename there so an operator can find
which report file was generated at which point in time.

Final note: this report is only supported on Linux systems. On macOS an
`info` line is emitted to remind its user that this only works on Linux.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>