fix: Tone down some spans, and instrument some others
Some of these were turned up to info for performance testing and are not necessary every day.  Some are duplicately nested, and we don't need all of them.  

I also added some spans for function execution, as this data would be incredibly helpful in debugging. However, this should still be a net-negative impact on the volume of spans.... famous last words...