fix(dal): Don't error when finding Attribute Values for components that are being deleted
To clean up frames during destroy actions, we check to see if the frame has a resource by finding the attribute value for `root/resource` by prop path. If there are multiple components for the same schema variant, and one has been deleted, but the snapshot hasn't been cleaned up, we'll stumble on the attribute value for that deleted component and returning an error, causing the component we're trying to delete to remain, leaving ghosted frames incorrectly.