Allow firecracker creation skip and fix config mismatch in veritech
This commit allows firecracker setup script creation to be skipped. This
is useful for a variety of reasons: getting a new rootfs, changing the
script itself, and more.

This commit also fixes a config mismatch in veritech where pinga's
concurrency limit will be derived for all global merged configurations.
Until we fix the "pool exhausted == NACK counting towards max failed
message count" problem, we cannot afford the concurrency limit to be
higher than the pool size. Now, we will be able to individual customize.

Why not pursue a dynamic solution where the concurrency limit is equal
to the pool size by default? A bit of magic lies in there as well as a
potential side effect to the other runtime strategies (local process and
docker). We are really only trying to target the use of concurrency limit
in production settings using firecracker, so renaming the variable will
suffice.

Signed-off-by: Nick Gerace <nick@systeminit.com>