chore: remove NodeIndex from all public WorkspaceSnapshot methods
In order to prepare for the split-graph to integrate with the DAL, we
need to remove all of the public methods on WorkspaceSnapshot that make
use of the petgraph NodeIndex type, since the SplitGraph will not expose
the NodeIndex type externally.  This is also a quality of life refactor,
which simplifies some "warty" usage patterns (where we were converting
between Ulid and NodeIndex, book-keeping work that the snapshot is
capable of doing internally).