feat(dal,web): Enable users to select which connections they want to create if Autoconnect detects too many options
This PR modifies the 'Autoconnect' logic and adds a user experience to accomodate.  Currently, when a user triggers autoconnect, if we find that there are multiple components/output sockets that could replace a manually set value, we do nothing and say 'No connections found'.  With this PR, we will accumulate the possible connections that *could* be made, and display those options so that the user can decide for themselves without having to manually go figure out where the options are in their workspace.

The UI also helps ensure the user doesn't accidentally do the wrong thing by:
- If the input socket in question in multi-arity, there are two 'flavors' of connections that can work. 
1. There could be an output socket somewhere that has that whole array (called "bulk" mode). There could be multiple output sockets like this. 
2. There could be many output sockets, that each drive a single entry in the array (called "individual" mode). There could be many options for a single entry as well. 

In any case, we want the user to know what's allowed, but we don't want them to accidentally create connections for the individual entries and leave one off, so we only let connections be created if all individual entries have corresponding connections. 

TESTING
I added more integration tests here around this logic, so if the feature picks up speed, we're less likely to regress! 

Manually tested by creating various permutations of optionality for both multi and single array sockets + matches, and seeing the connections get created correctly.