Disable tokio watchdog in si-service apps
This commit disables our tokio watchdog in si-service apps. This was
useful in earlier verideath investigation sessions, but we no longer
need it. Not only that, but the data returned was not entirely helpful
nor could we verify its accuracy. Given that this is spawned with every
backend service and sends INFO level logs to Honeycomb, we have decided
to remove it from "si-service". This will also slightly speed up compile
times for all services due to its removal from the "si-service" crate.

Signed-off-by: Nick Gerace <nick@systeminit.com>