fix(dal): Calculate Action Dependencies using DFS to capture transitive dependencies between components
Before this change, we were calculating the order for actions based only on the components with actions enqueued and their direct dependencies. This meant if you had Component A (with action) ---edge to --> Component B (no action) -- edge to --> Component C (with action), we would not detect that Action C should depend on Action A.  

With this change we now build a more full tree of component dependencies, and use petgraph's depth_first_search to find the closest dependency for a given component's action, which allows us to capture these transitive dependencies. 

TESTING
I added an integration test covering both forwards and backwards transitive dependencies (forwards applies to create/update/manual actions, backwards applies to destroy actions)