chore(firecracker): re-enable vsock telemetry forwarding
This change re-enables the stream forwarder in Cyclone and the listener
in Veritech, both of which help are used to forward OpenTelemetry data
out of a Firecracker VM.

Prior to disabling these tasks, it was unclear whether the spawned tokio
tasks would naturally terminate on their own, or whether they would
continue to run even as more Firecracker instances were spawing in Pool
Noodle's pool. This change attempts to tie the lifetime of these
forwarding tasks to the lifetime of the Firecracker instance (that is
the termination of the VM should also trigger the shutdown of the
forwarding task).

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>