feat: add optional tokio runtime worker thread core affinity support
This change adds a new optional capability in the `lib/si-runtime` crate
to setup a Tokio runtime that spawns its threads with a CPU core
affinity assignment. This way it's possible to run the main
multi-threaded runtime on specific CPU cores--that is the worker and
blocking pool threads.

An additional change is made in the `lib/si-service` crate to expose an
optional `rt::block_on_with_core_affinity()` function that takes a list
of `CoreId`s which is used when pinning the runtime threads.

Finally, Veritech's CLI args are augmented to include a
`--tokio-cpu-cores` option (and corresponding `SI_TOKIO_CPU_CORES`
environment variable) that will configure the runtime to pin to the
desired CPU cores.

For example running Veritech with:

```
veritech --tokio-cpu-cores 0-3
```

would configure the Tokio runtime to run only on cores 0,1,2,3. Without
this option, the default is typically a thread-per-core setup as before.

Future work may enable this capability for all other services.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>