feat(si-split-graph,dal): integrate si-split-graph
Integrates the "split graph" with the DAL.

Some functionality that was previously on the workspace snapshot, but
did not rely on internal implementation details and had no need of
`self` was moved off. For example, the DependentValueRoots and the
InputSocket::new code was moved off the snapshot, so that it could be
shared between legacy and split snapshots.

Adds the column `snapshot_kind` to the workpaces table to indicate
whether the snapshot is a "legacy" or "split".

Some functionality is not yet implemented for the split snapshot:
  1. cycle detection is implemented but is not correct.
  2. Approval flow is not yet implemented.
  3. The prop type trait needs to be re-implemented for the split
     snapshot.

Split graph code should only be executed if the workspace is marked as
split. The most important thing to review in this PR is that it does not
regress existing functionality and that we will not execute split
snapshot code on accident.