Update index if reference kinds do not match
For the new frontend architecture, this commit updates the index if
reference kinds found in the existing index do not match what is
currently available. This is useful when new types are added to the
system, but an existing system needs a new index to know that they
exist.

This takes about ~200ms when it happens, which we should keep an eye on
when equivalent functionality moves to "edda".

In the case of a rolling update with multiple SDF instances, it is
possible for both SDF instances to update the index. This is unlikely
given that a tenant will likely be "speaking" to one SDF and that this
is a stop-gap solution until "edda" takes on this behavior. However, it
is worth calling out.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>