Finish edda logic for real world use
This commit finishes "edda", a service written by @fnichol that is the
"keeper of the materialize views". All materialized view generation has
been moved out of the `dal` (and `sdf` and `rebaser` by extension) and
into `edda-server`. Using naxum and NATS KV, "edda" will process
materialized view generation requests for a given change set and publish
them accordingly. This reduces load on our other services while ensuring
we send quick updates to materialized views to the frontend. As part of
these changes, there is a new LayerDb table for "ChangeBatches", which
are used to "send" a batch of changes from the rebaser to edda.

Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>