Move actions list to norse stack
This commit moves the "list_actions" functionality to the norse stack.
It also brings in action prototype views into the fold from the
"AssetActionsDetails" component.

As a result of these changes, the new "mead-hall" directory contains
the beginning of flagged, bifrost-powered components that should be
guarded by feature flag. As things move over to the new paradigm, we do
not want to infect existing components. Rather, we want to feature flag
which component is used solely at the template level, if possible.

One of the materialized views from these changes involves a
multi-trigger approach where the entity kind is "root", but additional
checks need to be performed to know the minimal set of MVs to build. The
"build_mv!" macro has been changed to accomodate this.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>