Return better error for de-synced change set apply
This commit returns a user-friendly error when applying a change set
with a de-synced local state (spicedb, users table in PG, graph
requirements, etc.).

Why not solve the de-sync issue? This is not easily reproducible (words,
oh do not bite me) and would require a non-trivial amount of
investigation into syncing auth with local state.

Why use a toast? The current UX involves the change set button to
remain "clickable" and we use the "success" state to imply that. On the
"InsetApprovalModal", the space can get crowded with several reviewers
without a further look at its overall shape. Finally, for the standard
apply modal (used mostly when in a single user workspace), the current
UX involves the modal immediately closing due to the length of an apply
call to produce a response.

Is there a way to detect the de-synced state and ignore the error?
Without consistent reproduction, not really. Furthermore, the approval
logic is correct; it's just what data it queries that is out of date.
Getting this logic mangled with state detection may create more
trouble than it is trying to solve.

Signed-off-by: Nick Gerace <nick@systeminit.com>