Add base contents for job-specific UI
First and foremost, I want to be very clear that @jobelenus wrote all
of this. Wendy and I mainly polished it up.

This commit adds the "big boom, foundational pieces and solves some
problems" for the new UI. In addition, it showcases a README with
lines of achievement for the new UI... all written by @jobelenus.

This commit also adds a feature flag for the new UI and uses router
re-direction for gating the feature.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Co-authored-by: Wendy Bujalski <wendy@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>