Only build MVs where the MV ID is immediate
This commit ensures that we only build MVs where the MV ID is the same
as the trigger's change ID or is the change set ID. This ensures that
we do not run into situations where deleted nodes cannot be found when
"hopping" to the desired MV ID from a change ID.

As a result of these changes, this is the first commit where we have
removed an existing MV. It is no longer in use. We are not concerned
about backwards compatibility here since it is not deserialized
anywhere.

This commit also fixes the issue where the icons for action prototypes
were all question marks.

Signed-off-by: Nick Gerace <nick@systeminit.com>