feat(luminork): Ability to get a component by Id, find by name, and list all Component Ids for a workspace
This adds some initial functionality for listing component Ids, getting a component by Id, and finding a component by name. 

I also added Workpace Id and ChangeSet Id to the params for the swagger UI, mimicing what we had for the jwt, because I was sick of having to copy them in each time across each route.  Claude helped here, we needed to inject these as global parameters so that we have access to them, doing this in the interceptor didn't work as the OpenAI spec validates the request before it gets there and the validation was failing when workspace Id and change set Id were empty. 

I'm also experimenting with versioned API types in luminork itself, we can rip this into a separate crate once we get further and see the shapes of these things - we'll have to see how this feels.

Expect these shapes to change as we build out the API, but the scaffold is there to build on. 

Tested by executing the routes and seeing them work using the swagger-ui!