Add new UI survival guide and relieve macro builds
This commit adds a "newhotness" survival guide for those working on the
new UI, building MVs, working with edda, and more.

This commit also moves the "build_mv!" macro to a new crate called
"dal-materialized-views-macros" that only "edda-server" uses. This
ensures that changes to the "build_mv!" macro do not result in a rebuild
of the "dal". Not only that, but it ensures that the only crate that
can build MVs is "edda-server".

Finally, this commit renames "dal-newhotness" to
"dal-materialized-views" for better clarity.

Signed-off-by: Nick Gerace <nick@systeminit.com>