Restructure AttributeTree MV to be keyed on ComponentId
The prior version of this created a MV per `AttributeValue` across the
entire change set. This could be... a very large number of MV instances.

Rather than creating hundreds of thousands of MVs for moderately sized
workspaces & change sets, we now collate the `AttributeValue`
information into a single MV per Component. While this does mean that we
will rebuild the MV to see if it has changed when there have been
unrelated changes to the Component, it also means that we will always do
this once per Component at most, and are able to cache more information
used in the MV across all `AttributeValue`s in the Component.

Co-authored-by: Nick Gerace <nick@systeminit.com>