Reactivity Fixes & View list filter on explore
First we made the context we're providing into a computed structure.
Then Heimdall returned a computed structure out of the `useMakeKey` fn, with the standard `useXXX` pattern.
That key needs to be reactive, so that when the changeset changes, the tan stack query re-fires.

Adding `ViewComponentList` Mv where membership is based on geometry, even though we dont use the geo

Co-authored-by: Wendy <wendy@systeminit.com>