Add MVs for incoming socket-to-socket connections
This change adds MVs for incoming socket-to-socket connections. The
MVs have a "Beta" suffix for easy invalidations once we test all four
kinds of connections needed in the new UI.

As a consequence of these changes, the macro(s) used for MVs have been
adjusted to handle enums with struct variants.

For the record, @fnichol architected the vast majority of this. I just
wired it up to the frontend and drove it home. Go Fletcher!

Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>