Remove attribute tree MV
This change removes the attribute tree MV. While we have work to do to
make the new architecture more performant, this MV has been the likely
source of performance issues and is going to be replaced very soon
anyway.

We are explicitly keeping the frontend the same and are accepting
mjolnir 404s at the recommendation of @jobelenus.

Signed-off-by: Nick Gerace <nick@systeminit.com>